backend/requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.104.1
filelock==3.20.0
fsspec==2025.10.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openai==2.8.0
packaging==25.0
pillow==12.0.0
psutil==7.1.3
pydantic==2.12.4
pydantic-settings==2.0.3
pydantic_core==2.41.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.9.9
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
